Fix build for Java 8.
