+++ src/test/java/org/jilt/test/FullNameTest.java
@@ -3,8 +3,6 @@ package org.jilt.test;
 import org.jilt.test.data.nullable.FullName;
 import org.jilt.test.data.nullable.FullNameBuilder;
 import org.jilt.test.data.nullable.FullNameBuilders;
-import org.jilt.test.data.nullable.FullNameJSpecify;
-import org.jilt.test.data.nullable.FullNameJSpecifyBuilder;
 import org.junit.Test;
 
 import javax.annotation.Nullable;
@@ -39,18 +37,5 @@ public class FullNameTest {
         Method middleNameSetter = FullNameBuilders.MiddleName.class.getMethod("middleName", String.class);
 
         assertThat(middleNameSetter.getParameters()[0].getAnnotation(Nullable.class)).isNotNull();
-    }
-
-    @Test
-    public void jspecify_nullable_annotation_makes_attribute_optional() {
-        FullNameJSpecify value = FullNameJSpecifyBuilder.fullNameJSpecify()
-                .firstName("First")
-                // middleName is implicitly optional, because of @Nullable
-                .lastName("")
-                .build();
-
-        assertThat(value.firstName).isEqualTo("First");
-        assertThat(value.middleName).isNull();
-        assertThat(value.lastName).isEmpty();
     }
 }
