+++ src/test/java/org/jilt/test/data/nullable/JSpecifyFullName.java
@@ -5,7 +5,7 @@ import org.jilt.BuilderStyle;
 import org.jspecify.annotations.Nullable;
 
 @Builder(style = BuilderStyle.STAGED_PRESERVING_ORDER)
-public final class FullNameJSpecify {
+public final class JSpecifyFullName {
     public final String firstName;
 
     @Nullable
@@ -13,7 +13,7 @@ public final class FullNameJSpecify {
 
     public final String lastName;
 
-    public FullNameJSpecify(String firstName, @Nullable String middleName, String lastName) {
+    public JSpecifyFullName(String firstName, @Nullable String middleName, String lastName) {
         this.firstName = firstName;
         this.middleName = middleName;
         this.lastName = lastName;
